requirements.txt
python-telegram-bot>=20.0
schedule>=1.2.0
asyncio